Move contract management SYSCALLs to native
