+++ src/neo/SmartContract/ApplicationEngine.Contract.cs
@@ -1,5 +1,4 @@
 using Neo.Cryptography.ECC;
-using Neo.IO;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
 using Neo.SmartContract.Manifest;
@@ -12,11 +11,9 @@ namespace Neo.SmartContract
 {
     partial class ApplicationEngine
     {
-        public static readonly InteropDescriptor System_Contract_Create = Register("System.Contract.Create", nameof(CreateContract), 0, CallFlags.AllowModifyStates, false);
-        public static readonly InteropDescriptor System_Contract_Update = Register("System.Contract.Update", nameof(UpdateContract), 0, CallFlags.AllowModifyStates, false);
-        public static readonly InteropDescriptor System_Contract_Destroy = Register("System.Contract.Destroy", nameof(DestroyContract), 0_01000000, CallFlags.AllowModifyStates, false);
         public static readonly InteropDescriptor System_Contract_Call = Register("System.Contract.Call", nameof(CallContract), 0_01000000, CallFlags.AllowCall, false);
         public static readonly InteropDescriptor System_Contract_CallEx = Register("System.Contract.CallEx", nameof(CallContractEx), 0_01000000, CallFlags.AllowCall, false);
+        public static readonly InteropDescriptor System_Contract_CallNative = Register("System.Contract.CallNative", nameof(CallNativeContract), 0, CallFlags.None, false);
         public static readonly InteropDescriptor System_Contract_IsStandard = Register("System.Contract.IsStandard", nameof(IsStandardContract), 0_00030000, CallFlags.AllowStates, true);
         public static readonly InteropDescriptor System_Contract_GetCallFlags = Register("System.Contract.GetCallFlags", nameof(GetCallFlags), 0_00030000, CallFlags.None, false);
         /// <summary>
@@ -24,91 +21,8 @@ namespace Neo.SmartContract
         /// Warning: check first that input public key is valid, before creating the script.
         /// </summary>
         public static readonly InteropDescriptor System_Contract_CreateStandardAccount = Register("System.Contract.CreateStandardAccount", nameof(CreateStandardAccount), 0_00010000, CallFlags.None, true);
-
-        protected internal void CreateContract(byte[] nefFile, byte[] manifest)
-        {
-            if (!(ScriptContainer is Transaction tx))
-                throw new InvalidOperationException();
-            if (nefFile.Length == 0)
-                throw new ArgumentException($"Invalid NefFile Length: {nefFile.Length}");
-            if (manifest.Length == 0 || manifest.Length > ContractManifest.MaxLength)
-                throw new ArgumentException($"Invalid Manifest Length: {manifest.Length}");
-
-            AddGas(StoragePrice * (nefFile.Length + manifest.Length));
-
-            NefFile nef = nefFile.AsSerializable<NefFile>();
-            UInt160 hash = Helper.GetContractHash(tx.Sender, nef.Script);
-            ContractState contract = Snapshot.Contracts.TryGet(hash);
-            if (contract != null) throw new InvalidOperationException($"Contract Already Exists: {hash}");
-            contract = new ContractState
-            {
-                Id = Snapshot.ContractId.GetAndChange().NextId++,
-                UpdateCounter = 0,
-                Script = nef.Script,
-                Hash = hash,
-                Manifest = ContractManifest.Parse(manifest)
-            };
-
-            if (!contract.Manifest.IsValid(hash)) throw new InvalidOperationException($"Invalid Manifest Hash: {hash}");
-
-            Snapshot.Contracts.Add(hash, contract);
-
-            // We should push it onto the caller's stack.
-
-            Push(Convert(contract));
-
-            // Execute _deploy
-
-            ContractMethodDescriptor md = contract.Manifest.Abi.GetMethod("_deploy");
-            if (md != null)
-                CallContractInternal(contract, md, new Array(ReferenceCounter) { false }, CallFlags.All, ReturnTypeConvention.EnsureIsEmpty);
-        }
-
-        protected internal void UpdateContract(byte[] nefFile, byte[] manifest)
-        {
-            if (nefFile is null && manifest is null) throw new ArgumentException();
-
-            AddGas(StoragePrice * ((nefFile?.Length ?? 0) + (manifest?.Length ?? 0)));
-
-            var contract = Snapshot.Contracts.GetAndChange(CurrentScriptHash);
-            if (contract is null) throw new InvalidOperationException($"Updating Contract Does Not Exist: {CurrentScriptHash}");
-
-            if (nefFile != null)
-            {
-                if (nefFile.Length == 0)
-                    throw new ArgumentException($"Invalid NefFile Length: {nefFile.Length}");
-
-                NefFile nef = nefFile.AsSerializable<NefFile>();
-
-                // Update script
-                contract.Script = nef.Script;
-            }
-            if (manifest != null)
-            {
-                if (manifest.Length == 0 || manifest.Length > ContractManifest.MaxLength)
-                    throw new ArgumentException($"Invalid Manifest Length: {manifest.Length}");
-                contract.Manifest = ContractManifest.Parse(manifest);
-                if (!contract.Manifest.IsValid(contract.Hash))
-                    throw new InvalidOperationException($"Invalid Manifest Hash: {contract.Hash}");
-            }
-            contract.UpdateCounter++; // Increase update counter
-            if (nefFile != null)
-            {
-                ContractMethodDescriptor md = contract.Manifest.Abi.GetMethod("_deploy");
-                if (md != null)
-                    CallContractInternal(contract, md, new Array(ReferenceCounter) { true }, CallFlags.All, ReturnTypeConvention.EnsureIsEmpty);
-            }
-        }
-
-        protected internal void DestroyContract()
-        {
-            UInt160 hash = CurrentScriptHash;
-            ContractState contract = Snapshot.Contracts.TryGet(hash);
-            if (contract == null) return;
-            Snapshot.Contracts.Delete(hash);
-            foreach (var (key, _) in Snapshot.Storages.Find(BitConverter.GetBytes(contract.Id)))
-                Snapshot.Storages.Delete(key);
-        }
+        public static readonly InteropDescriptor System_Contract_NativeOnPersist = Register("System.Contract.NativeOnPersist", nameof(NativeOnPersist), 0, CallFlags.None, false);
+        public static readonly InteropDescriptor System_Contract_NativePostPersist = Register("System.Contract.NativePostPersist", nameof(NativePostPersist), 0, CallFlags.None, false);
 
         protected internal void CallContract(UInt160 contractHash, string method, Array args)
         {
@@ -172,6 +86,14 @@ namespace Neo.SmartContract
             }
         }
 
+        protected internal void CallNativeContract(string name)
+        {
+            NativeContract contract = NativeContract.GetContract(name);
+            if (contract is null || contract.ActiveBlockIndex > Snapshot.PersistingBlock.Index)
+                throw new InvalidOperationException();
+            contract.Invoke(this);
+        }
+
         protected internal bool IsStandardContract(UInt160 hash)
         {
             ContractState contract = Snapshot.Contracts.TryGet(hash);
@@ -207,6 +129,24 @@ namespace Neo.SmartContract
         protected internal UInt160 CreateStandardAccount(ECPoint pubKey)
         {
             return Contract.CreateSignatureRedeemScript(pubKey).ToScriptHash();
+        }
+
+        protected internal void NativeOnPersist()
+        {
+            if (Trigger != TriggerType.OnPersist)
+                throw new InvalidOperationException();
+            foreach (NativeContract contract in NativeContract.Contracts)
+                if (contract.ActiveBlockIndex <= Snapshot.PersistingBlock.Index)
+                    contract.OnPersist(this);
+        }
+
+        protected internal void NativePostPersist()
+        {
+            if (Trigger != TriggerType.PostPersist)
+                throw new InvalidOperationException();
+            foreach (NativeContract contract in NativeContract.Contracts)
+                if (contract.ActiveBlockIndex <= Snapshot.PersistingBlock.Index)
+                    contract.PostPersist(this);
         }
     }
 }
